requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
tavily-python
PyPDF2
#Authentication libraries
python-jose[cryptography] 
passlib[bcrypt] 
python-multipart